1. 删除两个无用按钮 2. 修复托管模式被自己的反作弊杀的bug

+++ YuEzTools/UI/TaskLevel.cs
@@ -49,7 +49,7 @@ public static class TaskLevel
             }
             num -= disconnectedCount;
             
-            float num2 = (float)instance.CompletedTasks / (float)instance.TotalTasks * 100;
+            double num2 = (float)instance.CompletedTasks / (float)instance.TotalTasks * 100;
 
             switch (GameManager.Instance.LogicOptions.GetTaskBarMode())
             {
